Updated Ticket Model to link to TicketUpdate and Customer
Ticket has a OneToMany relationship with TicketUpdate

There is a JsonManagedReference annotation on the List<TicketUpdate>
update property

This is going to be serialized into Json based on the TicketUpdate
property

There is a ManyToOne relationship to Customer
There can be multiple Tickets attached to a Customer.

JsonBackReference is attached to Customer customer property. This will
not be referenced in Json.

JsonProperties have been updated