Created Ticket-View Template
1. Imported the ticket model as a JSON object into the HTML doc

Why:

a. This ticket model contains all the information for viewing and
processing data

2. Dynamically created id's for the ticketUpdates

Why:
a. Need to be able to retrieve the list elements by id for script
processing.

b. The text inside these id's will be updated and sent through a form

3. Made p tags with id selectors contentEditable = false

Why:

a. This set's the document up to make the content editable utilizing
javascript